refactor(adapter-hooks): deno
fixes #210
