+++ packages/adapter-hooks/deps.js
@@ -1,0 +1,3 @@
+
+export * as R from 'https://cdn.skypack.dev/ramda@^0.27.1'
+export { default as crocks } from 'https://cdn.skypack.dev/crocks@^0.12.4'
