+++ packages/adapter-hooks/dev_deps.js
@@ -1,0 +1,3 @@
+
+export { assert, assertEquals, assertObjectMatch } from 'https://deno.land/std@0.98.0/testing/asserts.ts'
+export { spy, resolves } from 'https://deno.land/x/mock@v0.9.5/mod.ts'
