+++ packages/adapter-hooks/dev_deps_lock.json
@@ -1,0 +1,24 @@
+{
+  "https://deno.land/std@0.93.0/async/delay.ts": "9de1d8d07d1927767ab7f82434b883f3d8294fb19cad819691a2ad81a728cf3d",
+  "https://deno.land/std@0.98.0/fmt/colors.ts": "db22b314a2ae9430ae7460ce005e0a7130e23ae1c999157e3bb77cf55800f7e4",
+  "https://deno.land/std@0.98.0/testing/_diff.ts": "961eaf6d9f5b0a8556c9d835bbc6fa74f5addd7d3b02728ba7936ff93364f7a3",
+  "https://deno.land/std@0.98.0/testing/asserts.ts": "289b6ccf3b422ffb311cbf56996960d6530f8d8c0a7281803764c5ec672f9f9c",
+  "https://deno.land/x/collections@v0.11.0/common.ts": "34e8367e3696c3f872ae417d7c421fa908a5a2125a1c4cb259f7dee9561a7096",
+  "https://deno.land/x/collections@v0.11.0/comparators.ts": "08563340dbb0051f032bacdcf854bcabd13d607d2e8cb1889826417419df89d0",
+  "https://deno.land/x/collections@v0.11.0/trees/bs_node.ts": "854d39f6d60cdcb47e1183f0fa67091e6bad59dd2b13252a8b38b1b37269fa67",
+  "https://deno.land/x/collections@v0.11.0/trees/bs_tree.ts": "694605e10a9f56caf8468c5eea06d60a8d81521dee75d4524a5f790b1ee713b5",
+  "https://deno.land/x/collections@v0.11.0/trees/rb_node.ts": "e5eecc211933140e0535fc371a4990bd5ac889b00136c96a4a610dbbf78d9ecb",
+  "https://deno.land/x/collections@v0.11.0/trees/rb_tree.ts": "e1e5f4e26bc255ebf41d72b498a2af903af69f0391276a4a0eac6d46fc43f942",
+  "https://deno.land/x/collections@v0.11.0/vector.ts": "23cb91087cc89ce9a1e10954336484b537a44bd786e21babc310ae85cb7ad52d",
+  "https://deno.land/x/mixins@v0.7.2/apply.ts": "dad7095324f5ce23693a0bc0eb3238f230c0ed2160ea8c285f3773ff7c76dcb0",
+  "https://deno.land/x/mock@v0.9.5/callbacks.ts": "610924901137b6a8ee2f5801d06a48af18da84c9486dd4775ff301de8a8d2b29",
+  "https://deno.land/x/mock@v0.9.5/deps/std/async/delay.ts": "b0855e5e208bcc08a6f7cb4debfc3ed408a4a3f1bc00ce36235481b94613f018",
+  "https://deno.land/x/mock@v0.9.5/deps/udibo/collections/comparators.ts": "c780b51a4fcdac4d506bf254d54702fdd46051bde9367fc819bd72cea041ac29",
+  "https://deno.land/x/mock@v0.9.5/deps/udibo/collections/trees/rb_tree.ts": "24839f7b1f66291d10d6640cb0051143d9b7effbd1e74271c3b564bb31a5b977",
+  "https://deno.land/x/mock@v0.9.5/deps/udibo/collections/vector.ts": "93285b7af952652e9733795c5677666e978e279ef0b7dae6013e9e017c022b30",
+  "https://deno.land/x/mock@v0.9.5/deps/udibo/mixins/apply.ts": "d2446714efc056d2c6a373dbf7a1a1db236991cae60b673db6e17d5c4d1bb8b7",
+  "https://deno.land/x/mock@v0.9.5/mod.ts": "9b44e3cbe9955ecf9f05b06cb9e8a47294bd84f2d33e0b85a8f5b4a0d740e0aa",
+  "https://deno.land/x/mock@v0.9.5/spy.ts": "c6c66b124f1c0b86f479e66e1daf687ef03aff068ec776ff61a3bf25ff3da5d0",
+  "https://deno.land/x/mock@v0.9.5/stub.ts": "762cb86de92038e17dfd1af25429036fa7aadecbafdb39fa98edfa01e20951eb",
+  "https://deno.land/x/mock@v0.9.5/time.ts": "7c6d42af7577c4da461f9581bf171f6e5cb7a22a243ba87ee4ed3c4f1922c2e4"
+}
