--- packages/adapter-hooks/index_test.js
@@ -1,8 +1,0 @@
-const test = require('tape')
-const hooksAdapter = require('./index.js')
-
-test('call hooks', async t => {
-  const adapter = hooksAdapter().link()()
-  t.ok(adapter.call)
-  t.end()
-})
