+++ packages/adapter-hooks/package.json
@@ -3,19 +3,5 @@
   "version": "1.0.4",
   "homepage": "https://github.com/hyper63/hyper63#readme",
   "repository": "https://github.com/hyper63/hyper63",
-  "license": "Apache-2.0",
-  "main": "index.js",
-  "scripts": {
-    "test": "tape *_test.js"
-  },
-  "dependencies": {
-    "@vercel/fetch": "^6.1.0",
-    "crocks": "^0.12.4",
-    "node-fetch": "^2.6.1",
-    "ramda": "^0.27.1"
-  },
-  "devDependencies": {
-    "fetch-mock": "^9.10.7",
-    "tape": "^5.0.1"
-  }
+  "license": "Apache-2.0"
 }
